+++ crates/sncast/src/response/utils/class_hash.rs
@@ -1,5 +1,4 @@
 use crate::response::cast_message::SncastCommandMessage;
-use crate::response::{cast_message::SncastMessage, command::CommandResponse};
 use conversions::padded_felt::PaddedFelt;
 use conversions::{serde::serialize::CairoSerialize, string::IntoPaddedHexStr};
 use foundry_ui::styling;
@@ -10,15 +9,10 @@ pub struct ClassHashResponse {
     pub class_hash: PaddedFelt,
 }
 
-impl CommandResponse for ClassHashResponse {}
-
-impl SncastCommandMessage for SncastMessage<ClassHashResponse> {
+impl SncastCommandMessage for ClassHashResponse {
     fn text(&self) -> String {
         styling::OutputBuilder::new()
-            .field(
-                "Class Hash",
-                &self.command_response.class_hash.into_padded_hex_str(),
-            )
+            .field("Class Hash", &self.class_hash.into_padded_hex_str())
             .build()
     }
 }