+++ crates/sncast/src/response/utils/serialize.rs
@@ -1,6 +1,4 @@
 use crate::response::cast_message::SncastCommandMessage;
-use crate::response::cast_message::SncastMessage;
-use crate::response::command::CommandResponse;
 use conversions::serde::serialize::CairoSerialize;
 use foundry_ui::styling;
 use serde::{Deserialize, Serialize};
@@ -11,11 +9,9 @@ pub struct SerializeResponse {
     pub calldata: Vec<Felt>,
 }
 
-impl CommandResponse for SerializeResponse {}
-
-impl SncastCommandMessage for SncastMessage<SerializeResponse> {
+impl SncastCommandMessage for SerializeResponse {
     fn text(&self) -> String {
-        let calldata = format!("{:?}", &self.command_response.calldata);
+        let calldata = format!("{:?}", &self.calldata);
 
         styling::OutputBuilder::new()
             .field("Calldata", &calldata)