+++ crates/sncast/tests/integration/lib_tests.rs
@@ -4,10 +4,10 @@ use crate::helpers::constants::{
 use crate::helpers::fixtures::create_test_provider;
 
 use camino::Utf8PathBuf;
-use foundry_ui::UI;
 use shared::rpc::{get_rpc_version, is_expected_version};
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
+use sncast::response::ui::UI;
 use sncast::{check_if_legacy_contract, get_account, get_provider};
 use starknet_rust::accounts::Account;
 use starknet_rust::macros::felt;